# docs/api-spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/tickets": {
      "get": {
        "summary": "Get Tickets",
        "operationId": "get_tickets_tickets_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TicketItemResponse"
                  },
                  "title": "Response Get Tickets Tickets Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Ticket",
        "operationId": "create_ticket_tickets_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketItemResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tickets/{ticket_id}": {
      "get": {
        "summary": "Get Ticket By Id Route",
        "operationId": "get_ticket_by_id_route_tickets__ticket_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Ticket Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketItemResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Del Ticket",
        "operationId": "del_ticket_tickets__ticket_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Ticket Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tickets/{ticket_id}/assign": {
      "patch": {
        "summary": "Assign",
        "operationId": "assign_tickets__ticket_id__assign_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Ticket Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketItemResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tickets/{ticket_id}/start": {
      "patch": {
        "summary": "Start Maintenance Route",
        "operationId": "start_maintenance_route_tickets__ticket_id__start_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Ticket Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceItemResponse"
                }
              }
            }
          },
          "200": {
            "description": "Idempotent resume: a maintenance already existed for this ticket",
            "content": {
              "application/json": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tickets/{ticket_id}/cancel": {
      "patch": {
        "summary": "Cancel",
        "operationId": "cancel_tickets__ticket_id__cancel_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Ticket Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancellationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketItemResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tickets/{ticket_id}/addwkd": {
      "patch": {
        "summary": "Create Add Wkd",
        "operationId": "create_add_wkd_tickets__ticket_id__addwkd_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Ticket Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddWkdRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketItemResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddWkdRequest": {
        "properties": {
          "operation_percentage": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Operation Percentage"
          },
          "market_temperature": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Market Temperature"
          },
          "operation_damage": {
            "type": "boolean",
            "title": "Operation Damage"
          },
          "completed": {
            "type": "boolean",
            "title": "Completed"
          },
          "observations_wkd": {
            "type": "string",
            "title": "Observations Wkd"
          }
        },
        "type": "object",
        "required": [
          "operation_percentage",
          "market_temperature",
          "operation_damage",
          "completed",
          "observations_wkd"
        ],
        "title": "AddWkdRequest"
      },
      "AssignRequest": {
        "properties": {
          "technician_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Technician Id"
          }
        },
        "type": "object",
        "title": "AssignRequest"
      },
...
      "CancellationRequest": {
        "properties": {
          "cancellation_reason": {
            "type": "string",
            "title": "Cancellation Reason"
          }
        },
        "type": "object",
        "required": [
          "cancellation_reason"
        ],
        "title": "CancellationRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MaintenanceItemResponse": {
        "properties": {
          "maintenance_id": {
            "type": "string",
            "format": "uuid7",
            "title": "Maintenance Id"
          },
          "ticket_id": {
            "type": "integer",
            "title": "Ticket Id"
          },
          "maintenance_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maintenance Date"
          },
          "maintenance_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maintenance Description"
          },
          "spares": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SpareOut"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spares",
            "default": []
          },
          "technicians": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/TechnicianOut"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Technicians",
            "default": []
          },
          "initial_photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Initial Photo Url"
          },
          "pdf_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pdf Url"
          },
          "photos": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/PhotoOut"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photos",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "maintenance_id",
          "ticket_id"
        ],
        "title": "MaintenanceItemResponse"
      },
...
      "PhotoOut": {
        "properties": {
          "photo_id": {
            "type": "integer",
            "title": "Photo Id"
          },
          "photo_url": {
            "type": "string",
            "title": "Photo Url"
          }
        },
        "type": "object",
        "required": [
          "photo_id",
          "photo_url"
        ],
        "title": "PhotoOut"
      },
...
      "SpareOut": {
        "properties": {
          "spare_id": {
            "type": "integer",
            "title": "Spare Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "price": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Price"
          },
          "qty": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Qty"
          }
        },
        "type": "object",
        "required": [
          "spare_id",
          "name",
          "price",
          "qty"
        ],
        "title": "SpareOut"
      },
...
      "TechnicianOut": {
        "properties": {
          "technician_id": {
            "type": "integer",
            "title": "Technician Id"
          },
          "technician_name": {
            "type": "string",
            "title": "Technician Name"
          },
          "start_hour": {
            "type": "string",
            "format": "time",
            "title": "Start Hour"
          },
          "end_hour": {
            "type": "string",
            "format": "time",
            "title": "End Hour"
          }
        },
        "type": "object",
        "required": [
          "technician_id",
          "technician_name",
          "start_hour",
          "end_hour"
        ],
        "title": "TechnicianOut"
      },
...
      "TicketCreate": {
        "properties": {
          "ticket_id": {
            "type": "integer",
            "title": "Ticket Id"
          },
          "ticket_date": {
            "type": "string",
            "title": "Ticket Date"
          },
          "ticket_description": {
            "type": "string",
            "title": "Ticket Description"
          },
          "priority": {
            "type": "string",
            "title": "Priority"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "market_id": {
            "type": "integer",
            "title": "Market Id"
          },
          "equipment_id": {
            "type": "integer",
            "title": "Equipment Id"
          }
        },
        "type": "object",
        "required": [
          "ticket_id",
          "ticket_date",
          "ticket_description",
          "priority",
          "status",
          "market_id",
          "equipment_id"
        ],
        "title": "TicketCreate"
      },
      "TicketItemResponse": {
        "properties": {
          "ticket_id": {
            "type": "integer",
            "title": "Ticket Id"
          },
          "ticket_date": {
            "type": "string",
            "format": "date",
            "title": "Ticket Date"
          },
          "ticket_description": {
            "type": "string",
            "title": "Ticket Description"
          },
          "priority": {
            "type": "string",
            "title": "Priority"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "market_name": {
            "type": "string",
            "title": "Market Name"
          },
          "market_city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market City"
          },
          "equipment_name": {
            "type": "string",
            "title": "Equipment Name"
          },
          "assigned_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned Name"
          }
        },
        "type": "object",
        "required": [
          "ticket_id",
          "ticket_date",
          "ticket_description",
          "priority",
          "status",
          "market_name",
          "equipment_name"
        ],
        "title": "TicketItemResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}